+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 `2021.02.23`
 
 - feat: add `pull_request_target` trigger.
+- feat: add `translate-title` `translate-body`.
 
 ## v1.0.0